BROKEN COMMIT: All code from PNGReader.java has been copied over; no more code from the jazzlib project needs conversion.
...however, running WZ now silently fails (but doesn't exit...) The code needs to be verified. I'll do that later today, if possible.

+++ trunk/win32_source/WordBuilder.cpp
@@ -53,8 +53,8 @@ BOOL WordBuilder::moveRight(int amt) {
 
 	//Any change?
 	int newAmt = currSelectedID + amt;
-	if (newAmt >= possibleWords.size())
-		newAmt = possibleWords.size()-1;
+	if (newAmt >= (int)possibleWords.size())
+		newAmt = (int)possibleWords.size()-1;
 	else if (newAmt < 0)
 		newAmt = 0;
 	if (newAmt == currSelectedID)
